+++ crates/holochain_types/src/app.rs
@@ -53,10 +53,10 @@ pub enum CoordinatorSource {
 }
 
 #[derive(Clone, Debug, serde::Serialize, serde::Deserialize)]
-/// The instructions on how to update coordinators for a dna file.
+/// The instructions on how to update coordinators for a cell.
 pub struct UpdateCoordinatorsPayload {
-    /// The hash of the dna to swap coordinators for.
-    pub dna_hash: DnaHash,
+    /// The cell id of the cell to swap coordinators for.
+    pub cell_id: CellId,
     /// Where to find the coordinators.
     pub source: CoordinatorSource,
 }